refactor: split permission kind from access helper
